fix: move wheel listener to scroll render controller
